requirements.txt
googletrans == 4.0.0-rc1
discord.py == 1.6.0
emoji == 1.2.0
